requirements.txt
requests==2.31.0
psycopg==3.1.18
python-dotenv==1.0.0
selenium==4.15.2